add: mga 2 polygon deployment
